Restructure datasets

+++ artifacts/scripts/run-JSONAlgebra.sh
@@ -1,8 +1,28 @@
 #!/bin/bash
+
+while [[ $# -gt 0 ]]; do
+  case $1 in
+    -i|--input)
+      input=$2
+      shift
+      shift
+      ;;
+    -*|--*|*)
+      echo "Unknown option $1"
+      exit 1
+      ;;
+  esac
+done
+
 cd ${HOME}/JSONAlgebra
 export MAVEN_OPTS="-Xmx10240m"
 
 run_experiment() {
+    if [ ! -d "${HOME}/JSONAlgebra/JsonSchema_To_Algebra/expDataset/${1}" ]; then
+      echo "Dataset ${input} not found."
+      return
+    fi
+
     mkdir -p ${HOME}/results/${1//\//-}/
     rm JsonSchema_To_Algebra/expDataset/${1}/results/[0-9]*_results.csv 2> /dev/null
     mvn exec:java -Dexec.mainClass="it.unipi.di.tesiFalleniLandi.JsonSchema_to_Algebra.MassiveTesting.MainClassV2" \
@@ -10,6 +30,13 @@ run_experiment() {
             2> ${HOME}/results/${1//\//-}/${1//\//-}-err.log
     cp ${HOME}/JSONAlgebra/JsonSchema_To_Algebra/expDataset/${1}/results/[0-9]*_results.csv ${HOME}/results/${1//\//-}/results.csv 2> /dev/null
 }
+
+# If specified, run experiments only on the given input dataset. Otherwise run experiments on all default datasets
+if [ -n "$input" ];
+  then
+    run_experiment $input
+    exit 0
+fi
 
 echo "Running experiments on Containment dataset..."
 run_experiment containment/sat
@@ -21,7 +48,14 @@ run_experiment handwritten/unsat
 
 
 echo "Running experiments on Snowplow dataset..."
-run_experiment snowplow
+run_experiment snowplow/ours
+# Move results from snowplow-ours to snowplow
+(
+    cd ${HOME}/results
+    mkdir snowplow 2> /dev/null
+    mv snowplow-ours/jsongenerator_results snowplow/jsongenerator_results
+    rm -r snowplow-ours
+)
 # Copy Snowplow results to charts
 cp ${HOME}/results/snowplow/results.csv ${HOME}/charts/data/snowplow/results.csv
 
@@ -55,7 +89,7 @@ run_experiment kubernetes/unsat
     mkdir kubernetes 2> /dev/null
     awk '(NR == 1) || (FNR > 1)' kubernetes-sat/results.csv \
         kubernetes-unsat/results.csv > kubernetes/results.csv
-	# Copy Kubernetes results to charts
-	cp ${HOME}/results/kubernetes/results.csv ${HOME}/charts/data/kubernetes/results.csv
-	rm -r kubernetes
+    # Copy Kubernetes results to charts
+    cp ${HOME}/results/kubernetes/results.csv ${HOME}/charts/data/kubernetes/results.csv
+    rm -r kubernetes
 )
