+++ artifacts/scripts/run-jsongenerator.sh
@@ -1,15 +1,40 @@
 #!/bin/bash
+
+while [[ $# -gt 0 ]]; do
+  case $1 in
+    -i|--input)
+      input=$2
+      shift
+      shift
+      ;;
+    -*|--*|*)
+      echo "Unknown option $1"
+      exit 1
+      ;;
+  esac
+done
+
 cd ${HOME}/jsongenerator
 
-gradle build
-
 run_experiment() {
+    if [ ! -d "${HOME}/JSONAlgebra/JsonSchema_To_Algebra/expDataset/${1}" ]; then
+      echo "Dataset ${input} not found."
+      return
+    fi
+
     rm ${HOME}/JSONAlgebra/JsonSchema_To_Algebra/expDataset/${1}/results/jsongenerator_* 2> /dev/null
     gradle run -Pdata="['/home/repro/JSONAlgebra/JsonSchema_To_Algebra/expDataset/${1}/']"
     mkdir -p ${HOME}/results/${1//\//-}/
     cp ${HOME}/JSONAlgebra/JsonSchema_To_Algebra/expDataset/${1}/results/jsongenerator_*_results.csv \
         ${HOME}/results/${1//\//-}/jsongenerator_results.csv 2> /dev/null
 }
+
+# If specified, run experiments only on the given input dataset. Otherwise run experiments on all default datasets
+if [ -n "$input" ];
+  then
+    run_experiment $input
+    exit 0
+fi
 
 # Containment Data Set
 run_experiment containment/sat
@@ -26,24 +51,22 @@ run_experiment github/unsat
 # Move results from github-sat-dg to github-sat
 (
     cd ${HOME}/results
-	mkdir github-sat 2> /dev/null
-	mv github-sat-dg/jsongenerator_results.csv github-sat/jsongenerator_results.csv
+    mkdir github-sat 2> /dev/null
+    mv github-sat-dg/jsongenerator_results.csv github-sat/jsongenerator_results.csv
     rm -r github-sat-dg
 )
 
 run_experiment kubernetes/sat
 run_experiment kubernetes/unsat
 
-run_experiment snowplow
 run_experiment snowplow/dg
-# Combine results from snowplow and snowplow-dg
+
+# Move results from snowplow-dg to snowplow
 (
     cd ${HOME}/results
-	mv snowplow/jsongenerator_results.csv snowplow/jsongenerator_results_part.csv
-    awk '(NR == 1) || (FNR > 1)' snowplow/jsongenerator_results_part.csv \
-        snowplow-dg/jsongenerator_results.csv > snowplow/jsongenerator_results.csv
-	rm snowplow/jsongenerator_results_part.csv
-	rm -r snowplow-dg
+    mkdir snowplow 2> /dev/null
+    mv snowplow-dg/jsongenerator_results.csv snowplow/jsongenerator_results.csv
+    rm -r snowplow-dg
 )
 
 run_experiment wp
